requirements.txt
flask
pyyaml
biopython
pandas
matplotlib